requirements.txt
Flask==3.0.3
Pillow==10.3.0
numpy==2.0.0